Dedupe generator script helpers and format repo count consistently
scripts/lib/s3-assets.mjs and scripts/update-github-repo-dates.mjs
each had their own copy of collectContentFiles (with slightly
different, silently divergent extension/ignore rules) and an
identical yamlQuote helper. Extract both into a shared
scripts/lib/generator-utils.mjs that both scripts import, with each
caller keeping its own extension/ignore-set arguments so behavior is
unchanged. Also format the logged repo count with toLocaleString for
consistency with generate-s3-asset-versions.mjs.
